+++ graphene/contrib/django/filter/__init__.py
@@ -1,0 +1,3 @@
+from .fields import DjangoFilterConnectionField
+from .filterset import GrapheneFilterSet, GlobalIDFilter
+from .resolvers import FilterConnectionResolver
